+++ src/client/menu/server-info.ts
@@ -1,6 +1,7 @@
 import { RemoteServerConnectionSettings } from '../../server/remote/remote-server'
 import { DEFAULT_HTTP_PORT } from '../../net/web-server'
 import { PhysicsServerConnectionSettings } from '../../server/physics/physics-server'
+import { Opts } from '../../options'
 
 interface ServerDetailsBase {
     title: string
@@ -30,21 +31,49 @@ export interface NetServerInfoPhysics {
     details: ServerDetailsPhysics
 }
 
+export const serverListDefault: NetServerInfoRemote[] = [
+    {
+        connection: {
+            type: 'socket',
+            host: 'crosscode.krypek.cc',
+            port: 443,
+        },
+    },
+    {
+        connection: {
+            type: 'socket',
+            host: '127.0.0.1',
+            port: DEFAULT_HTTP_PORT,
+        },
+    },
+]
 export function getServerListInfo(): NetServerInfoRemote[] {
-    return [
-        {
-            connection: {
-                type: 'socket',
-                host: '127.0.0.1',
-                port: DEFAULT_HTTP_PORT,
-            },
-        },
-        {
-            connection: {
-                type: 'socket',
-                host: 'crosscode.krypek.cc',
-                port: 443,
-            },
-        },
-    ]
+    return [...Opts.serverList]
+}
+
+export function addServerListEntry(entry: NetServerInfoRemote) {
+    const servers = getServerListInfo()
+    servers.push(entry)
+    Opts.serverList = servers
+}
+export function removeServerListEntry(index: number) {
+    const servers = getServerListInfo()
+    servers.splice(index, 1)
+    Opts.serverList = servers
+}
+
+export function moveServerEntry(index: number, dir: -1 | 1): boolean {
+    const servers = getServerListInfo()
+    const newIndex = Math.min(servers.length - 1, Math.max(0, index + dir))
+    if (index == newIndex) return false
+    ;[servers[index], servers[newIndex]] = [servers[newIndex], servers[index]]
+    Opts.serverList = servers
+
+    return true
+}
+
+export function replaceServerEntry(index: number, entry: NetServerInfoRemote) {
+    const servers = getServerListInfo()
+    servers[index] = entry
+    Opts.serverList = servers
 }